Created script to help with commenting SAS
I needed a way to quickly make a header for a SAS macro file and be
able to use easier tags than <b> for bold, etc.  So I created this
script which adds a header to a macro file or replaces quick tags such
as `` with the html proper (<code>).